+++ Pmdg737/PublishChannel.ps1
@@ -1,0 +1,7 @@
+$currentDir = $pwd.Path
+$pathChannel = Resolve-Path "PMDG.737.json"
+
+cd ..\dist
+.\AddChannelToRepo.ps1 $pathChannel
+
+cd $currentDir